requirements.txt
torch
transformers
datasets
accelerate
evaluate
matplotlib
jupyter
scikit-learn
networkx
tqdm
numpy
scipy
python-dotenv
pandas